Advert fix test

+++ EcommerceBundle/Controller/AdvertController.php
@@ -7,11 +7,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-use EcommerceBundle\Entity\Advert;
-use EcommerceBundle\Form\AdvertType;
-use EcommerceBundle\Form\AdvertEditType;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use CoreBundle\Entity\PostalCode;
+use EcommerceBundle\Entity\Advert;
 
 /**
  * Advert controller.
@@ -63,45 +60,34 @@ class AdvertController extends Controller
     /**
      * Creates a new Advert entity.
      *
-     * @Route("/")
-     * @Method("POST")
-     * @Template("EcommerceBundle:Advert:new.html.twig")
+     * @Route("/new")
+     * @Method({"GET", "POST"})
+     * @Template()
      */
-    public function createAction(Request $request)
+    public function newAction(Request $request)
     {
-
         $entity = new Advert();
         $form = $this->createCreateForm($entity);
         $form->handleRequest($request);
-        
-        if ($form->isValid()) {
+
+        if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             
             //add postal codes
-            $data = $request->request->get('ecommercebundle_advert');
-
-            $postalCodes = explode(',', $data['codes']);
+            $data = $request->request->get('advert');
             $days = $data['days'];
             $sectionCount = count($data['located']);
-            
-            foreach ($postalCodes as $value) {
-                $postalCodeEntity = $em->getRepository('CoreBundle:PostalCode')->findOneByPostalCode($value);
-                if($postalCodeEntity instanceof PostalCode){
-                    $entity->addPostalCode($postalCodeEntity);
-                }
-            }
-            
             $em->persist($entity);
             $em->flush();
             
             //calculate total
             $params = $this->getParameter('core');
             $unitPrice = $params['ecommerce']['advert_unit_price'];
-            $quantity = count($postalCodes);
+            $quantity = 1;
             $discount = 0;
             $subtotal = (($unitPrice * $quantity* $days) * $sectionCount ) - $discount;
             $totalPrice = ($subtotal * 0.21) + $subtotal;
-             
+            
             $creditCard = $form->getNormData()->getCreditCard();
             //proccess sale
             $checkoutManager = $this->get('checkout_manager');
@@ -126,24 +112,19 @@ class AdvertController extends Controller
 
             if($answer->redirectUrl == '/response-ok'){
                 $this->get('session')->getFlashBag()->add('success', 'advert.created');
-                
-                $user = $this->container->get('security.token_storage')->getToken()->getUser();  
-
                 return $this->redirect($this->generateUrl('ecommerce_advert_show', array('id' => $entity->getId())));
-                
             }else{
                 $this->get('session')->getFlashBag()->add('success', 'advert.cancel');
                 return $this->redirect($this->generateUrl('ecommerce_advert_index'));
             }
-            
         }
 
         return array(
             'entity' => $entity,
-            'form'   => $form->createView(),
+            'form' => $form->createView(),
         );
     }
-
+    
     /**
      * Creates a form to create a Advert entity.
      *
@@ -159,31 +140,11 @@ class AdvertController extends Controller
             $formConfig['actor'] = $user;
         }
         $form = $this->createForm('EcommerceBundle\Form\AdvertType', $entity, array_merge($formConfig, array(
-            'action' => $this->generateUrl('ecommerce_advert_create'),
+            'action' => $this->generateUrl('ecommerce_advert_new'),
             'method' => 'POST',
         )));
 
-//        $form->add('submit', 'submit', array('label' => 'Create'));
-
         return $form;
-    }
-
-    /**
-     * Displays a form to create a new Advert entity.
-     *
-     * @Route("/new")
-     * @Method("GET")
-     * @Template()
-     */
-    public function newAction()
-    {
-        $entity = new Advert();
-        $form   = $this->createCreateForm($entity);
-
-        return array(
-            'entity' => $entity,
-            'form'   => $form->createView(),
-        );
     }
 
     /**
@@ -193,20 +154,12 @@ class AdvertController extends Controller
      * @Method("GET")
      * @Template()
      */
-    public function showAction($id)
+    public function showAction(Advert $advert)
     {
-        $em = $this->getDoctrine()->getManager();
-
-        $entity = $em->getRepository('EcommerceBundle:Advert')->find($id);
-
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Advert entity.');
-        }
-
-        $deleteForm = $this->createDeleteForm($id);
+        $deleteForm = $this->createDeleteForm($advert);
 
         return array(
-            'entity'      => $entity,
+            'entity' => $advert,
             'delete_form' => $deleteForm->createView(),
         );
     }
@@ -215,28 +168,38 @@ class AdvertController extends Controller
      * Displays a form to edit an existing Advert entity.
      *
      * @Route("/{id}/edit")
-     * @Method("GET")
+     * @Method({"GET", "POST"})
      * @Template()
      */
-    public function editAction($id)
+    public function editAction(Request $request, Advert $advert)
     {
-        $em = $this->getDoctrine()->getManager();
+        
+        $deleteForm = $this->createDeleteForm($advert);
+        $editForm = $this->createEditForm($advert);
+        $editForm->handleRequest($request);
 
-        $entity = $em->getRepository('EcommerceBundle:Advert')->find($id);
-
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Advert entity.');
+        if ($editForm->isSubmitted() && $editForm->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            
+            if($advert->getRemoveImage()){
+                $advert->setImage(null);
+            }
+            
+            $em->persist($advert);
+            $em->flush();
+            
+            $this->get('session')->getFlashBag()->add('success', 'advert.edited');
+            
+            return $this->redirectToRoute('ecommerce_advert_edit', array('id' => $advert->getId()));
         }
 
-        $editForm = $this->createEditForm($entity);
-        $deleteForm = $this->createDeleteForm($id);
-
         return array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
+            'entity' => $advert,
+            'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         );
     }
+     
 
     /**
     * Creates a form to edit a Advert entity.
@@ -247,102 +210,62 @@ class AdvertController extends Controller
     */
     private function createEditForm(Advert $entity)
     {
+        
         $formConfig = array();
         $user = $this->container->get('security.token_storage')->getToken()->getUser();  
         if ($user->isGranted('ROLE_ADMIN') || $user->isGranted('ROLE_SUPER_ADMIN')) {
             $formConfig['admin'] = true;
         }
-        $form = $this->createForm(new AdvertEditType($formConfig), $entity, array(
-            'action' => $this->generateUrl('ecommerce_advert_update', array('id' => $entity->getId())),
-            'method' => 'PUT',
-        ));
-
-//        $form->add('submit', 'submit', array('label' => 'Update'));
-
+        $form = $this->createForm('EcommerceBundle\Form\AdvertEditType', $entity, array_merge($formConfig, array(
+            'action' => $this->generateUrl('ecommerce_advert_edit', array('id' => $entity->getId())),
+            'method' => 'POST',
+        )));
+        
         return $form;
     }
+    
     /**
-     * Edits an existing Advert entity.
-     *
-     * @Route("/{id}")
-     * @Method("PUT")
-     * @Template("EcommerceBundle:Advert:edit.html.twig")
-     */
-    public function updateAction(Request $request, $id)
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $entity = $em->getRepository('EcommerceBundle:Advert')->find($id);
-
-        if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Advert entity.');
-        }
-
-        $deleteForm = $this->createDeleteForm($id);
-        $editForm = $this->createEditForm($entity);
-        $editForm->handleRequest($request);
-
-        if ($editForm->isValid()) {
-            $em->flush();
-
-            $this->get('session')->getFlashBag()->add('success', 'advert.edited');
-            
-            return $this->redirect($this->generateUrl('ecommerce_advert_edit', array('id' => $id)));
-        }
-
-        return array(
-            'entity'      => $entity,
-            'edit_form'   => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        );
-    }
-    /**
-     * Deletes a Advert entity.
+     * Deletes a Category entity.
      *
      * @Route("/{id}")
      * @Method("DELETE")
      */
-    public function deleteAction(Request $request, $id)
+    public function deleteAction(Request $request, Advert $advert)
     {
-        $form = $this->createDeleteForm($id);
+        $form = $this->createDeleteForm($advert);
         $form->handleRequest($request);
 
-        if ($form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $entity = $em->getRepository('EcommerceBundle:Advert')->find($id);
-
-            if (!$entity) {
-                throw $this->createNotFoundException('Unable to find Advert entity.');
-            }
+        if ($form->isSubmitted() && $form->isValid()) {
             
+            $em = $this->getDoctrine()->getManager();
             //remove product purchase, transaction and invoice
-            $purchases = $em->getRepository('EcommerceBundle:ProductPurchase')->findByAdvert($id);
+            $purchases = $em->getRepository('EcommerceBundle:ProductPurchase')->findByAdvert($advert);
             foreach ($purchases as $purchase) {
                 $em->remove($purchase);
             }
-            $em->remove($entity);
+            $em->remove($advert);
             $em->flush();
             
             $this->get('session')->getFlashBag()->add('info', 'advert.deleted');
         }
 
-        return $this->redirect($this->generateUrl('ecommerce_advert_index'));
+        return $this->redirectToRoute('ecommerce_advert_index');
     }
 
-    /**
-     * Creates a form to delete a Advert entity by id.
+   /**
+     * Creates a form to delete a Advert entity.
      *
-     * @param mixed $id The entity id
+     * @param Advert $advert The Advert entity
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createDeleteForm($id)
+    private function createDeleteForm(Advert $advert)
     {
         return $this->createFormBuilder()
-            ->setAction($this->generateUrl('ecommerce_advert_delete', array('id' => $id)))
+            ->setAction($this->generateUrl('ecommerce_advert_delete', array('id' => $advert->getId())))
             ->setMethod('DELETE')
-//            ->add('submit', 'submit', array('label' => 'Delete'))
             ->getForm()
         ;
     }
+   
 }
